Implement loading animation for SprintListActivity
Remove padding from sprint list to match other lists
design.

Add throwIfNull(...) for all list activities that get
serializable extras.

Restructure sprint list activity.

Fix bug in SprintOverviewActivity that causes the app
to crash when opening the activity. A null object was
added to the issueList, which would cause the adapter
to crash when attempting to sort the list.